codex P1/P2 対応 + SHOULD 補完 (Issue #24 PR#31)
## codex P1: _is_pid_alive_windows の ctypes restype/argtypes 明示

Win64 で `kernel32.OpenProcess` の戻り値 HANDLE はポインタ幅 (64bit) だが、
ctypes default の `restype=c_int` (32bit signed) のままだと高位ビットが
立った handle で truncate + sign-extend が起き、`GetExitCodeProcess` /
`CloseHandle` に誤った handle を渡し得る。`_win_kernel32()` ヘルパで
`wintypes.HANDLE/DWORD/BOOL` を明示し、bound symbol を一度だけ設定して
キャッシュ。launcher の < 100ms 起動 budget も維持。

## codex P2: server.json の lock + compare-and-delete を共有モジュール化

`hooks/launch_dashboard.py:_remove_stale_server_json` は素 read+unlink で
TOCTOU race の元のまま、`dashboard/server.py:remove_server_json` だけが
lock + compare-and-delete だった。launcher hook が ms 単位で連続発火する
シナリオで、launcher A が pid=X を読んだ後に新サーバーが pid=Y で
上書きすると、A が誤って新レジストリを削除する race が残存。

新規 `server_registry.py` (stdlib のみ) に `_file_lock` / `_lock_path_for`
/ `_pid_matches` / `write_server_json` / `remove_server_json` を集約し、
`dashboard/server.py` と `hooks/launch_dashboard.py` の両方から共有。
launcher の cleanup も `server_registry.remove_server_json(expected_pid=pid)`
経由で同一 lock + compare-and-delete を経由するようになり、multi-hook
race window で TOCTOU race が再発しない。

`dashboard/server.py` は既存テスト互換のため変数 binding 経由で再 export
(`mod._file_lock` 等は引き続き動作)。ただし monkeypatch は本元の
`server_registry` を対象にする必要があるため `tests/test_dashboard_live.py`
の 3 件を `server_registry._lock_fd` 経由に書き換え。

## SHOULD 補完: rescan_transcripts.py --help の Windows パス例

Issue #24 最新コメント (G) で「PR#31 で対応」宣言済みだが未実装だった。
`--transcripts-dir` の help に `~/.claude/projects on POSIX,
%USERPROFILE%\\.claude\\projects on Windows` を明記。

## 検証

- pytest 345/345 PASS
- pylint 10.00/10 (warning 0)
